v2.1.0

+++ src/mosaicmpi/cnmf.py
@@ -487,7 +487,7 @@ class cNMF():
         # with usages fixed and TPM as the input matrix
         tpm = sc.read(self.paths['tpm'])
         tpm_stats = utils.load_df_from_npz(self.paths['tpm_stats'])
-        norm_usages = norm_usages.fillna(0).astype(np.float64)  # replaces NaN usages for samples that have 0 HVG counts
+        norm_usages = norm_usages.fillna(0).astype(np.float32)  # replaces NaN usages for samples that have 0 HVG counts
         spectra_tpm = self.refit_spectra(tpm.X, norm_usages)
         spectra_tpm = pd.DataFrame(spectra_tpm, index=rf_usages.columns, columns=tpm.var.index)
         spectra_tpm = spectra_tpm.div(spectra_tpm.sum(axis=1), axis=0) * 1e6
